Add files via upload
added a python project.. 
Test the Fine Logic:

Add a book (e.g., "Python Basics").

Issue it to yourself.

Tip: To see the "Fine" work immediately, find the line date_now = datetime.now().strftime("%Y-%m-%d") in the code and change it to date_now = "2023-01-01" (a past date). Then run it and try to return the book. It will show a huge fine!